requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
cobble==0.1.4
Django==5.1.6
django-allauth==65.4.1
django-filter==25.1
djangorestframework==3.15.2
emoji==2.14.1
faiss-cpu==1.11.0
filelock==3.17.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-language==2.17.0
google-generativeai==0.8.5
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
mammoth==1.9.0
Markdown==3.8
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.3
PyPDF2==3.0.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
selenium-stealth==1.0.6
sentence-transformers==4.1.0
setuptools==75.8.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stanza==1.10.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
schedule==1.2.0
lxml==4.9.3
python-docx==1.1.2
weasyprint==62.3
reportlab==4.0.9
requests-html==0.10.0
pyppeteer==1.0.2
undetected-chromedriver==3.5.5
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
vis-network==1.1.0
python-magic==0.4.27
nltk==3.8.1
celery[redis]==5.3.4
redis==5.0.1
flower==2.0.1
httpx==0.27.0
fastmcp==1.1.0
markitdown==0.0.1a6
